chore: provide a canonical "failing" `BlackBoxSolver`
