+++ acvm-repo/blackbox_solver/src/curve_specific_solver.rs
@@ -1,0 +1,79 @@
+use acir::{BlackBoxFunc, FieldElement};
+
+use crate::BlackBoxResolutionError;
+
+/// This component will generate outputs for Blackbox function calls where the underlying [`acir::BlackBoxFunc`]
+/// doesn't have a canonical Rust implementation.
+///
+/// Returns an [`BlackBoxResolutionError`] if the backend does not support the given [`acir::BlackBoxFunc`].
+pub trait BlackBoxFunctionSolver {
+    fn schnorr_verify(
+        &self,
+        public_key_x: &FieldElement,
+        public_key_y: &FieldElement,
+        signature: &[u8],
+        message: &[u8],
+    ) -> Result<bool, BlackBoxResolutionError>;
+    fn pedersen_commitment(
+        &self,
+        inputs: &[FieldElement],
+        domain_separator: u32,
+    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError>;
+    fn pedersen_hash(
+        &self,
+        inputs: &[FieldElement],
+        domain_separator: u32,
+    ) -> Result<FieldElement, BlackBoxResolutionError>;
+    fn fixed_base_scalar_mul(
+        &self,
+        low: &FieldElement,
+        high: &FieldElement,
+    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError>;
+}
+
+pub struct StubbedSolver;
+
+impl BlackBoxFunctionSolver for StubbedSolver {
+    fn schnorr_verify(
+        &self,
+        _public_key_x: &FieldElement,
+        _public_key_y: &FieldElement,
+        _signature: &[u8],
+        _message: &[u8],
+    ) -> Result<bool, BlackBoxResolutionError> {
+        Err(BlackBoxResolutionError::Failed(
+            BlackBoxFunc::SchnorrVerify,
+            "SchnorrVerify is not supported".to_string(),
+        ))
+    }
+    fn pedersen_commitment(
+        &self,
+        _inputs: &[FieldElement],
+        _domain_separator: u32,
+    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError> {
+        Err(BlackBoxResolutionError::Failed(
+            BlackBoxFunc::PedersenCommitment,
+            "PedersenCommitment is not supported".to_string(),
+        ))
+    }
+    fn pedersen_hash(
+        &self,
+        _inputs: &[FieldElement],
+        _domain_separator: u32,
+    ) -> Result<FieldElement, BlackBoxResolutionError> {
+        Err(BlackBoxResolutionError::Failed(
+            BlackBoxFunc::PedersenHash,
+            "PedersenHash is not supported".to_string(),
+        ))
+    }
+    fn fixed_base_scalar_mul(
+        &self,
+        _low: &FieldElement,
+        _high: &FieldElement,
+    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError> {
+        Err(BlackBoxResolutionError::Failed(
+            BlackBoxFunc::FixedBaseScalarMul,
+            "FixedBaseScalarMul is not supported".to_string(),
+        ))
+    }
+}
