+++ tooling/lsp/src/solver.rs
@@ -40,45 +40,3 @@ impl BlackBoxFunctionSolver for WrapperSolver {
         self.0.pedersen_hash(inputs, domain_separator)
     }
 }
-
-// We also have a mocked implementation of the `BlackBoxFunctionSolver` trait for use in tests
-
-#[cfg(test)]
-pub(crate) struct MockBackend;
-
-#[cfg(test)]
-impl BlackBoxFunctionSolver for MockBackend {
-    fn schnorr_verify(
-        &self,
-        _public_key_x: &acvm::FieldElement,
-        _public_key_y: &acvm::FieldElement,
-        _signature: &[u8],
-        _message: &[u8],
-    ) -> Result<bool, acvm::BlackBoxResolutionError> {
-        unimplemented!()
-    }
-
-    fn pedersen_commitment(
-        &self,
-        _inputs: &[acvm::FieldElement],
-        _domain_separator: u32,
-    ) -> Result<(acvm::FieldElement, acvm::FieldElement), acvm::BlackBoxResolutionError> {
-        unimplemented!()
-    }
-
-    fn fixed_base_scalar_mul(
-        &self,
-        _low: &acvm::FieldElement,
-        _high: &acvm::FieldElement,
-    ) -> Result<(acvm::FieldElement, acvm::FieldElement), acvm::BlackBoxResolutionError> {
-        unimplemented!()
-    }
-
-    fn pedersen_hash(
-        &self,
-        _inputs: &[acvm::FieldElement],
-        _domain_separator: u32,
-    ) -> Result<acvm::FieldElement, acvm::BlackBoxResolutionError> {
-        unimplemented!()
-    }
-}
